requirements.txt
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
pathspec==0.10.2
platformdirs==2.5.4
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typer==0.7.0
urllib3==1.26.12